requirements.txt
# ── requirements.txt ───────────────────────────
# ===== Runtime (生产/开发共用) =================
Django==5.2.3                  # 主框架
djangorestframework==3.16.0    # API
djangorestframework-simplejwt==5.3.1  # JWT
django-cors-headers==4.3.1     # CORS
django-filter==24.1            # DRF 过滤器
psycopg2-binary==2.9.9         # Postgres 驱动
gunicorn==21.2.0               # 部署 WSGI

# ===== Dev / CI Only ==========================
flake8==6.1.0                  # 代码规范
pytest==8.2.0                  # 单元测试
pytest-django==4.8.0           # Django 测试插件
pytest-cov==5.0.0              # 覆盖率
